feat(topbar): externalize logic, update layout and responsiveness
